Revert "Added lots of warning flags and adjusted code thereafter."
This reverts commit 5a22999892d241cc3a4a29e62de1af503a8e200c.

+++ lib/smooth/include/smooth/application/network/mqtt/MqttClient.h
@@ -102,7 +102,7 @@ namespace smooth
                         /// \return true if the message could be queued for delivery, otherwise false. A true value
                         /// does not mean it has been delivered.
                         bool
-                        publish(const std::string& topic, const uint8_t* data, size_t length, mqtt::QoS qos, bool retain);
+                        publish(const std::string& topic, const uint8_t* data, int length, mqtt::QoS qos, bool retain);
 
                         /// Subscribes to a topic.
                         /// \param topic The topic
