+++ lib/smooth/include/smooth/application/network/mqtt/state/MqttFsmConstants.h
@@ -5,7 +5,7 @@
 #pragma once
 
 // Set this to the size of the largest state
-const size_t MQTT_FSM_STATE_SIZE = 150;
+const int MQTT_FSM_STATE_SIZE = 150;
 
 const int MQTT_FSM_RECONNECT_TIMER_ID = 1;
 const int MQTT_FSM_KEEP_ALIVE_TIMER_ID = 2;
